# [hig-vanilla] Update Webpack config