--- a/packages/vanilla/.babelrc
+++ b/packages/vanilla/.babelrc
@@ -1,0 +1,8 @@
+{
+  "presets": [
+    "es2015",
+    "stage-2"
+  ],
+  "plugins": [],
+  "compact": false
+}