--- a/packages/vanilla/.babelrc
+++ b/packages/vanilla/.babelrc
@@ -3,6 +3,5 @@
     "es2015",
     "stage-2"
   ],
-  "plugins": [],
-  "compact": false
+  "plugins": []
 }