--- a/packages/vanilla/src/basics/form-elements/checkbox/checkbox.js
+++ b/packages/vanilla/src/basics/form-elements/checkbox/checkbox.js
@@ -1,6 +1,6 @@
 import Interface from 'interface.json';
 import Template from './checkbox.html';
-import InputButton from '../input-button/input-button';
+import InputButton from 'basics/form-elements/input-button/input-button';
 
 /**
  * Creates an Checkbox