--- a/packages/vanilla/src/basics/form-elements/password-field/password-field.js
+++ b/packages/vanilla/src/basics/form-elements/password-field/password-field.js
@@ -1,7 +1,7 @@
 import Template from './password-field.html';
 import Interface from 'interface.json';
 import Core from '_core.js';
-import TextField from '../text-field/text-field';
+import TextField from 'basics/form-elements/text-field/text-field';
 
 const textFieldMethods = [
   'setLabel',