--- a/packages/vanilla/src/basics/form-elements/radio-button/radio-button.js
+++ b/packages/vanilla/src/basics/form-elements/radio-button/radio-button.js
@@ -1,6 +1,6 @@
 import Interface from 'interface.json';
 import Template from './radio-button.html';
-import InputButton from '../input-button/input-button';
+import InputButton from 'basics/form-elements/input-button/input-button';
 
 
 // List all the fn you don't want to override so they can be defined in this prototype