--- a/packages/vanilla/src/components/button/button.js
+++ b/packages/vanilla/src/components/button/button.js
@@ -2,7 +2,7 @@ import Interface from 'interface.json';
 import Core from '_core.js';
 import './button.scss';
 import Template from './button.html';
-import Icon from '../../basics/icon/icon';
+import Icon from 'basics/icon/icon';
 
 const AvailableTypes = ['primary', 'secondary', 'flat'];
 const AvailableSizes = ['small', 'standard', 'large'];