--- a/packages/vanilla/src/components/global-nav/global-nav.js
+++ b/packages/vanilla/src/components/global-nav/global-nav.js
@@ -2,9 +2,9 @@ import Interface from 'interface.json';
 import Core from '_core.js';
 import './global-nav.scss';
 import Template from './global-nav.html';
-import SideNav from './side-nav/side-nav.js';
-import TopNav from './top-nav/top-nav.js';
-import SubNav from './sub-nav/sub-nav.js';
+import SideNav from 'components/global-nav/side-nav/side-nav';
+import TopNav from 'components/global-nav/top-nav/top-nav';
+import SubNav from 'components/global-nav/sub-nav/sub-nav';
 
 /**
  * Creates an GlobalNav