--- a/packages/vanilla/src/components/global-nav/top-nav/help/group/group.js
+++ b/packages/vanilla/src/components/global-nav/top-nav/help/group/group.js
@@ -2,7 +2,7 @@ import Interface from 'interface.json';
 import Core from '_core.js';
 import './group.scss';
 import Template from './group.html';
-import Option from '../option/option.js';
+import Option from 'components/global-nav/top-nav/help/option/option';
 
 /**
  * Creates a Group