--- a/packages/vanilla/src/components/global-nav/top-nav/profile/profile.js
+++ b/packages/vanilla/src/components/global-nav/top-nav/profile/profile.js
@@ -2,9 +2,9 @@ import Interface from 'interface.json';
 import Core from '_core.js';
 import './profile.scss';
 import Template from './profile.html';
-import Flyout from './../../../../basics/flyout/flyout';
-import ProfileImage from './_profile-image/profile-image';
-import ProfileFlyoutContent from './_profile-flyout-content/profile-flyout-content';
+import Flyout from 'basics/flyout/flyout';
+import ProfileImage from 'components/global-nav/top-nav/profile/_profile-image/profile-image';
+import ProfileFlyoutContent from 'components/global-nav/top-nav/profile/_profile-flyout-content/profile-flyout-content';
 /**
  * Creates an Profile
  *