--- a/packages/vanilla/src/components/global-nav/top-nav/project-account-switcher/_lists/lists.js
+++ b/packages/vanilla/src/components/global-nav/top-nav/project-account-switcher/_lists/lists.js
@@ -1,7 +1,7 @@
 import Interface from 'interface.json';
 import Core from '_core.js';
 import Template from './lists.html';
-import List from './_list/_list.js';
+import List from 'components/global-nav/top-nav/project-account-switcher/_lists/_list/_list';
 
 /**
  * Creates an Lists