--- a/packages/vanilla/src/components/global-nav/top-nav/project-account-switcher/project/project.js
+++ b/packages/vanilla/src/components/global-nav/top-nav/project-account-switcher/project/project.js
@@ -1,5 +1,5 @@
 import Interface from 'interface.json';
-import Item from '../_item/item';
+import Item from 'components/global-nav/top-nav/project-account-switcher/_item/item';
 
 /**
  * Creates an Project