--- a/packages/vanilla/src/components/global-nav/top-nav/search/search.js
+++ b/packages/vanilla/src/components/global-nav/top-nav/search/search.js
@@ -2,8 +2,8 @@ import Interface from 'interface.json';
 import Core from '_core.js';
 import './search.scss';
 import Template from './search.html';
-import Icon from '../../../../basics/icon/icon';
-import Option from '../../../../basics/form-elements/option/option';
+import Icon from 'basics/icon/icon';
+import Option from 'basics/form-elements/option/option';
 
 const OPEN_CLASS = 'hig__option__list--show';
 /**