--- a/packages/vanilla/src/components/global-nav/top-nav/shortcut/shortcut.js
+++ b/packages/vanilla/src/components/global-nav/top-nav/shortcut/shortcut.js
@@ -2,7 +2,7 @@ import Interface from 'interface.json';
 import Core from '_core.js';
 import './shortcut.scss';
 import Template from './shortcut.html';
-import Icon from '../../../../basics/icon/icon';
+import Icon from 'basics/icon/icon';
 
 /**
  * Creates an Shortcut