--- a/packages/vanilla/src/components/modal/modal.js
+++ b/packages/vanilla/src/components/modal/modal.js
@@ -2,8 +2,8 @@ import Interface from 'interface.json';
 import Core from '_core.js';
 import './modal.scss';
 import Template from './modal.html';
-import Button from '../button/button';
-import Icon from '../../basics/icon/icon.js';
+import Button from 'components/button/button';
+import Icon from 'basics/icon/icon.js';
 
 const AvailableStyles = ['standard', 'alternate'];