--- a/packages/vanilla/src/components/progress-ring/progress-ring-determinate.js
+++ b/packages/vanilla/src/components/progress-ring/progress-ring-determinate.js
@@ -1,4 +1,4 @@
-import CSSTransition from '../../helpers/js/css-transition';
+import CSSTransition from 'helpers/js/css-transition';
 
 const MAX_INCREASE_PER_MS = 1 / 1000;