--- a/packages/vanilla/src/components/progress-ring/progress-ring-indeterminate.js
+++ b/packages/vanilla/src/components/progress-ring/progress-ring-indeterminate.js
@@ -1,4 +1,4 @@
-import CSSTransition from '../../helpers/js/css-transition';
+import CSSTransition from 'helpers/js/css-transition';
 import svg from './progress-ring-s.svg';
 
 const CYCLE_DURATION = 1000;