--- a/packages/vanilla/src/components/table/table-head/slot-head-cell/slot-head-cell.js
+++ b/packages/vanilla/src/components/table/table-head/slot-head-cell/slot-head-cell.js
@@ -2,7 +2,7 @@ import Interface from 'interface.json';
 import Core from '_core.js';
 import './slot-head-cell.scss';
 import Template from './slot-head-cell.html';
-import frToPercentage from '../../../../helpers/js/_frtopercentage';
+import frToPercentage from 'helpers/js/_frtopercentage';
 
 /**
  * Creates an Table