--- a/packages/vanilla/src/components/table/table-row/table-row.js
+++ b/packages/vanilla/src/components/table/table-row/table-row.js
@@ -2,9 +2,9 @@ import Interface from 'interface.json';
 import Core from '_core.js';
 import './table-row.scss';
 import Template from './table-row.html';
-import TextCell from './text-cell/text-cell.js';
-import SlotCell from './slot-cell/slot-cell.js';
-import IconCell from './icon-cell/icon-cell.js';
+import TextCell from 'components/table/table-row/text-cell/text-cell';
+import SlotCell from 'components/table/table-row/slot-cell/slot-cell';
+import IconCell from 'components/table/table-row/icon-cell/icon-cell';
 
 /**
  * Creates an Table