--- a/packages/vanilla/webpack.dev.js
+++ b/packages/vanilla/webpack.dev.js
@@ -1,0 +1,6 @@
+const merge = require('webpack-merge');
+const common = require('./webpack.common.js');
+
+module.exports = merge(common, {
+  devtool: 'inline-source-map',
+});