--- a/packages/vanilla/webpack.prod.js
+++ b/packages/vanilla/webpack.prod.js
@@ -1,0 +1,9 @@
+const merge = require('webpack-merge');
+const common = require('./webpack.common.js');
+const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
+
+module.exports = merge(common, {
+  plugins: [
+    new OptimizeCssAssetsPlugin(),
+  ],
+});