… packages/vanilla/src/components/progress-ring/progress-ring.js
… packages/vanilla/webpack.config.js