# Goals

1. Customize webpack configs based on whether vanilla is being built for development or production.
2. Move babel config out of webpack config into its own config file
3. Migrate import paths away from relative paths to "semi-relative" paths (from `packages/vanilla/src`)

# Timing

👏 Faster builds for development:

```bash
# before
time npx webpack --config webpack.config.js
24.42s user 1.22s system 107% cpu 23.778 total

# after
time npx webpack --config webpack.dev.js
17.68s user 0.85s system 111% cpu 16.684 total
```