@aneyzberg 2017-12-06
Looks good. I'll let @nfiniteset take a look in case he has thoughts